Move specified color types into specified::color mod.
